set to int16

+++ pepper_variant/modules/python/DataStorePredict.py
@@ -71,7 +71,7 @@ class DataStore(object):
             self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
                                                       chunk_name_suffix, 'index')] = np.array(index, dtype=np.int32)
             self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
-                                                      chunk_name_suffix, 'base_predictions')] = np.array(base_predictions, dtype=np.int32)
+                                                      chunk_name_suffix, 'base_predictions')] = np.array(base_predictions, dtype=np.int16)
 
     def write_prediction_hp(self, contig, contig_start, contig_end, chunk_id, position, index, base_predictions_hp1, base_predictions_hp2):
         chunk_name_prefix = str(contig) + "-" + str(contig_start.item()) + "-" + str(contig_end.item())
@@ -100,8 +100,8 @@ class DataStore(object):
             self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
                                                       chunk_name_suffix, 'index')] = np.array(index, dtype=np.int32)
             self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
-                                                      chunk_name_suffix, 'base_predictions_hp1')] = base_predictions_hp1.astype(np.int32)
+                                                      chunk_name_suffix, 'base_predictions_hp1')] = base_predictions_hp1.astype(np.int16)
             self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
-                                                      chunk_name_suffix, 'base_predictions_hp2')] = base_predictions_hp2.astype(np.int32)
+                                                      chunk_name_suffix, 'base_predictions_hp2')] = base_predictions_hp2.astype(np.int16)
